Move tooltip to common controls
