+++ src/components/common/index.ts
@@ -16,3 +16,4 @@
 
 export * from './dropdown';
 export * from './select';
+export * from './tooltip';
