+++ src/components/common/tooltip/index.js
@@ -1,5 +1,4 @@
 import 'react-tippy/dist/tippy.css';
 import { Tooltip, withTooltip } from 'react-tippy';
 
-export default Tooltip;
-export { withTooltip };
+export { Tooltip, withTooltip };
